Changed types to float

+++ lib.cpp
@@ -6,23 +6,23 @@
 #include <iterator>
 #include <cmath>
 
-int plus(int a, int b) { return a + b; }
+float plus(float a, float b) { return a + b; }
 
-int minus(int a, int b) { return a - b; }
+float minus(float a, float b) { return a - b; }
 
-int multiply(int a, int b) { return a * b; }
+float multiply(float a, float b) { return a * b; }
 
-int divide(int a, int b) {
+float divide(float a, float b) {
     if(b == 0)
         throw std::invalid_argument("division by zero is not allowed");
     return a / b;
 }
-int ln(         int a) { return std::log(a); }
-int exponential(int a) { return std::exp(a); }
-int square(     int a) { return a * a; }
-int squareroot( int a) { return std::sqrt(a); }
+float ln(         float a) { return std::log(a); }
+float exponential(float a) { return std::exp(a); }
+float square(     float a) { return a * a; }
+float squareroot( float a) { return std::sqrt(a); }
 
-int pop_stack(stack_type &stack) {
+float pop_stack(stack_type &stack) {
     if(stack.empty())
         throw std::invalid_argument("invalid input");
     const auto ret{stack.top()};
@@ -36,14 +36,14 @@ void change_stack_state(const binary_int_function &operation, stack_type &stack)
     stack.push(operation(op_a, op_b));
 }
 void apply_binary_function(const binary_int_function &operation, stack_type &stack) {
-  const auto op_b{pop_stack(stack)};
-  const auto op_a{pop_stack(stack)};
-  stack.push(operation(op_a, op_b));
+    const auto op_b{pop_stack(stack)};
+    const auto op_a{pop_stack(stack)};
+    stack.push(operation(op_a, op_b));
 }
 
 void apply_unary_function(const unary_int_function &operation, stack_type &stack) {
-  const auto op_a{pop_stack(stack)};
-  stack.push(operation(op_a));
+    const auto op_a{pop_stack(stack)};
+    stack.push(operation(op_a));
 }
 
 std::vector<std::string> splitStringBySpace(const std::string& input) {
@@ -64,30 +64,30 @@ bool isStringNumber(const std::string& input) {
 
 void processOperator(const char op, stack_type &stack) {
     switch (op) {
-    case '+': change_stack_state(plus,     stack); break;
-    case '-': change_stack_state(minus,    stack); break;
-    case '*': change_stack_state(multiply, stack); break;
-    case '/': change_stack_state(divide,   stack); break;
+        case '+': change_stack_state(plus,     stack); break;
+        case '-': change_stack_state(minus,    stack); break;
+        case '*': change_stack_state(multiply, stack); break;
+        case '/': change_stack_state(divide,   stack); break;
     }
 }
 
-int evaluate(const std::string &s) {
-  stack_type stack;
+float evaluate(const std::string &s) {
+    stack_type stack;
 
-  for (const char c : s) {
-    switch (c) {
-    case '+': apply_binary_function(plus,     stack); break;
-    case '-': apply_binary_function(minus,    stack); break;
-    case '*': apply_binary_function(multiply, stack); break;
-    case '/': apply_binary_function(divide,   stack); break;
-    case 'l': apply_unary_function(ln, stack); break;
-    case 'e': apply_unary_function(exponential, stack); break;
-    case 's': apply_unary_function(square, stack); break;
-    case 'q': apply_unary_function(squareroot, stack); break;
-    case '0'...'9': stack.push(c - '0'); break;
-    default:
+    for (const char c : s) {
+        switch (c) {
+            case '+': apply_binary_function(plus,     stack); break;
+            case '-': apply_binary_function(minus,    stack); break;
+            case '*': apply_binary_function(multiply, stack); break;
+            case '/': apply_binary_function(divide,   stack); break;
+            case 'l': apply_unary_function(ln, stack); break;
+            case 'e': apply_unary_function(exponential, stack); break;
+            case 's': apply_unary_function(square, stack); break;
+            case 'q': apply_unary_function(squareroot, stack); break;
+            case '0'...'9': stack.push(c - '0'); break;
+            default:
                 throw std::invalid_argument("invalid input");
-      }
+        }
     }
     if(stack.empty())
         throw std::invalid_argument("invalid input");
